+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed minimum requirements to `COMPAS 2.x`.
 * Moved `RobotModel`, `ToolModel` and `Configuration` to the top-level package of this library.
 * Moved `URDF` parsing from core into this library.
+* Changed urdf parsing to use `compas.colors.Color` instead of a robot-specific color class.
 
 ### Removed